AI - Move bid purchase and place methods to new class to isolate old Strong AI code (no method implementation changes)
